lambda-app/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
mangum==0.17.0
Werkzeug==3.0.1
